+++ llvm/lib/IR/CMakeLists.txt
@@ -79,6 +79,7 @@ add_llvm_component_library(LLVMCore
   ValueSymbolTable.cpp
   VectorTypeUtils.cpp
   Verifier.cpp
+  VerifierAMDGPU.cpp
   VFABIDemangler.cpp
   RuntimeLibcalls.cpp
