+++ llvm/test/Verifier/callbr-intrinsic.ll
@@ -2,7 +2,7 @@
 
 declare void @llvm.amdgcn.kill(i1)
 
-; CHECK: Callbr amdgcn_kill only supports one indirect dest
+; CHECK: callbr amdgcn_kill only supports one indirect dest
 define void @test_callbr_intrinsic_indirect0(i1 %c) {
   callbr void @llvm.amdgcn.kill(i1 %c) to label %cont []
 kill:
@@ -11,7 +11,7 @@ cont:
   ret void
 }
 
-; CHECK-NEXT: Callbr amdgcn_kill only supports one indirect dest
+; CHECK-NEXT: callbr amdgcn_kill only supports one indirect dest
 define void @test_callbr_intrinsic_indirect2(i1 %c) {
   callbr void @llvm.amdgcn.kill(i1 %c) to label %cont [label %kill1, label %kill2]
 kill1:
@@ -22,7 +22,7 @@ cont:
   ret void
 }
 
-; CHECK-NEXT: Callbr amdgcn_kill indirect dest needs to be unreachable
+; CHECK-NEXT: callbr amdgcn_kill indirect dest needs to be unreachable
 define void @test_callbr_intrinsic_no_unreachable(i1 %c) {
   callbr void @llvm.amdgcn.kill(i1 %c) to label %cont [label %kill]
 kill:
@@ -31,7 +31,7 @@ cont:
   ret void
 }
 
-; CHECK-NEXT: Callbr currently only supports asm-goto and selected intrinsics
+; CHECK-NEXT: callbr currently only supports asm-goto and selected intrinsics
 declare i32 @llvm.amdgcn.workitem.id.x()
 define void @test_callbr_intrinsic_unsupported() {
   callbr i32 @llvm.amdgcn.workitem.id.x() to label %cont []
@@ -39,7 +39,7 @@ cont:
   ret void
 }
 
-; CHECK-NEXT: Callbr: indirect function / invalid signature
+; CHECK-NEXT: callbr: indirect function / invalid signature
 define void @test_callbr_intrinsic_wrong_signature(ptr %ptr) {
   %func = load ptr, ptr %ptr, align 8
   callbr void %func() to label %cont []
@@ -47,7 +47,7 @@ cont:
   ret void
 }
 
-; CHECK-NEXT: Callbr for intrinsics currently doesn't support operand bundles
+; CHECK-NEXT: callbr for intrinsics currently doesn't support operand bundles
 define void @test_callbr_intrinsic_no_operand_bundles(i1 %c) {
   callbr void @llvm.amdgcn.kill(i1 %c) [ "foo"(i1 %c) ] to label %cont [label %kill]
 kill:
