+++ llvm/utils/gn/secondary/llvm/lib/IR/BUILD.gn
@@ -95,5 +95,6 @@ static_library("IR") {
     "ValueSymbolTable.cpp",
     "VectorTypeUtils.cpp",
     "Verifier.cpp",
+    "VerifierAMDGPU.cpp",
   ]
 }
